PyTeapotPlus/requirements.txt
pygame-ce~=2.5.6
pyopengl~=3.1.10
pyopengl-accelerate~=3.1.10
PyWavefront~=1.3.3
pyglet<2

